refactor: max time by year parameterized test
